requirements.txt
# Flask==3.0.3
cloud-sql-python-connector==1.11.0
pg8000==1.31.2
SQLAlchemy==2.0.32
# vertexai==1.60.0
google-cloud-aiplatform==1.60.0
google-generativeai==0.7.2
uvicorn==0.30.6
pandas==2.0.3
numpy==1.26.4
fastapi==0.112.1